Create an ExposomeSet from data.frames

Takes the three tables of an exposome dataset and coerces them into a Exposome Set object on the study server

ds.loadExposome(
  exposures,
  phenotypes,
  exposures.idcol = "idcol",
  phenotypes.idcol = "idcol",
  description = NULL,
  description.expCol = "exposure",
  description.famCol = "family",
  exposures.asFactor = 5,
  warnings = FALSE,
  object_name = "exposome_set",
  datasources = NULL
)

Arguments

exposures

character Name of the exposures variable on the study server

phenotypes

character Name of the phenotypes variable on the study server

exposures.idcol

character (default "idcol") Name of the column in the Exposures file that contains the individuals ID

phenotypes.idcol

character (default "idcol") Name of the column in the Phenotypes file that contains the individuals ID

description

character Name of the description variable on the study server. If there is no description file input NULL, each exposure will be assigned the same family name as it's exposure name (default value).

description.expCol

character (default "exposure") Name of the column in the Description file that contains the Exposure names

description.famCol

character (default "family") Name of column where the family's name (per exposures) if found in file "description".

exposures.asFactor

numeric (default 5) The exposures with more than this number of unique items will be considered as "continuous" while the exposures with less or equal number of items will be considered as "factor".

warnings

bool (default FALSE) If TRUE shows useful information/warnings from the process of loading the exposome.

object_name

character (default "exposome_set") Name assigned to the Exposome Set on the study server

datasources

a list of DSConnection-class (default NULL) objects obtained after login

Value

This function does not have an output. It creates an Exposome Set object on the study server.

Examples