Draw boxplot of a family of an Exposome Set of the server side

Draw a boxplot for a single family or a mosaic of boxplots for all the families (only numeric families)

ds.plotFamily(
  x,
  family,
  group = NULL,
  group2 = NULL,
  scatter = FALSE,
  na.omit = TRUE,
  datasources = NULL
)

Arguments

x

character Name of the Exposome Set on the server side

family

character Name of the familty that will be drawn. "all" is to plot all the families on a single mosaic (no grouping is used for the mosaic plot).

group

character If set it displays the family grouped by the given phenotype

group2

character If set it displays the family grouped by the given phenotype

scatter

bool (default TRUE) If the family to be plotted is continuous, the samples will be shown

na.omit

bool (default TRUE) Do not show NA values

datasources

a list of DSConnection-class (default NULL) objects obtained after login

method

numeric either 1 or 2. If the user selects the deterministic method in the client side function the method.inticator is set to 1 while if the user selects the probabilistic method this argument is set to 2.

k

numeric the number of the nearest neghbours for which their centroid is calculated

noise

numeric the percentage of the initial variance that is used as the variance of

Value

The function returns a ggplot object when a single family is selected and returns NULL for the mosaic (it renders the plot on the current graphic device of the session)

Examples